+++ CHANGES.md
@@ -21,7 +21,7 @@ Changes:
 -
 
 Fixes:
--
+- Fix `texapp` skip cases with no opinion documents to prevent empty sub_opinions #1852
 
 ## 2.8.3 - 2026-03-13